+++ src/models/character/CharacterStory.ts
@@ -1,4 +1,4 @@
-import { Client } from '@/client'
+import { Client } from '@/client/Client'
 
 /**
  * Contains character background story content and narrative elements
