+++ src/models/weapon/WeaponAscension.ts
@@ -1,8 +1,8 @@
-import { Client } from '@/client'
-import { AssetNotFoundError } from '@/errors'
+import { Client } from '@/client/Client'
+import { AssetNotFoundError } from '@/errors/assets/AssetNotFoundError'
 import { StatProperty } from '@/models/StatProperty'
-import { createPromoteLevelSchema } from '@/schemas'
-import { ValidationHelper } from '@/utils/validation'
+import { createPromoteLevelSchema } from '@/schemas/createPromoteLevelSchema'
+import { ValidationHelper } from '@/utils/validation/ValidationHelper'
 
 /**
  * Handles weapon enhancement data including promote levels, costs, and stat boosts
